backend/requirements.txt
fastapi
uvicorn[standard]
pytesseract
transformers
firebase-admin
python-multipart
Pillow
